Dangerous typecasts
Former-commit-id: 17950d09867a6af5a7a789f3c3425829d005c211

+++ source/core/body.cpp
@@ -169,7 +169,7 @@ void TBody::CleanHeatLayer()
 	if (!this) return;
 	if (!HeatLayer) return;
 
-	for (long i=0; i<List->size(); i++)
+	for (size_t i=0; i<List->size(); i++)
 	{
 		HeatLayer[i]=0;
 	}
